+++ src/js/pyodide.ts
@@ -10,7 +10,7 @@ import {
   resolvePath,
 } from "./compat";
 
-import { createModule, setStandardStreams, setHomeDirectory } from "./module";
+import { createModule, setHomeDirectory } from "./module";
 import { initializeNativeFS } from "./nativefs";
 import version from "./version";
 
@@ -247,8 +247,8 @@ export async function loadPyodide(
      * The home directory which Pyodide will use inside virtual file system. Default: "/home/pyodide"
      */
     homedir?: string;
-
-    /** Load the full Python standard library.
+    /**
+     * Load the full Python standard library.
      * Setting this to false excludes unvendored modules from the standard library.
      * Default: false
      */
@@ -295,6 +295,8 @@ export async function loadPyodide(
   );
 
   const Module = createModule();
+  Module.print = config.stdout;
+  Module.printErr = config.stderr;
   Module.preRun.push(() => {
     for (const mount of config._node_mounts) {
       Module.FS.mkdirTree(mount);
@@ -306,7 +308,6 @@ export async function loadPyodide(
   const API: any = { config };
   Module.API = API;
 
-  setStandardStreams(Module, config.stdin, config.stdout, config.stderr);
   setHomeDirectory(Module, config.homedir);
 
   const moduleLoaded = new Promise((r) => (Module.postRun = r));
@@ -343,13 +344,12 @@ If you updated the Pyodide version, make sure you also updated the 'indexURL' pa
 `,
     );
   }
-
-  initializeNativeFS(Module);
-
   // Disable further loading of Emscripten file_packager stuff.
   Module.locateFile = (path: string) => {
     throw new Error("Didn't expect to load any more file_packager files!");
   };
+
+  initializeNativeFS(Module);
 
   const pyodide_py_tar = await pyodide_py_tar_promise;
   unpackPyodidePy(Module, pyodide_py_tar);
@@ -375,5 +375,6 @@ If you updated the Pyodide version, make sure you also updated the 'indexURL' pa
   if (config.fullStdLib) {
     await pyodide.loadPackage(API._pyodide._importhook.UNVENDORED_STDLIBS);
   }
+  API.initializeStreams(config.stdin, config.stdout, config.stderr);
   return pyodide;
 }